Allow float values in altitude_in_meters
- Convert Spatial.altitude_in_meters into FloatField
- Normalize altitude (round to significant digits) for diff comparison
- Validate that altitude is a valid float in V2->V3 conversion
- Support marking invalid fields multiple times for same object

+++ src/apps/core/models/concepts.py
@@ -173,7 +173,7 @@ class Spatial(AbstractBaseModel):
     )
     full_address = models.CharField(max_length=512, blank=True, null=True)
     geographic_name = models.CharField(max_length=512, blank=True, null=True)
-    altitude_in_meters = models.IntegerField(
+    altitude_in_meters = models.FloatField(
         blank=True,
         null=True,
         help_text="The altitude of the geographical area (meters from WGS84 reference)",
